server/requirements.txt
flask
transformers 
torch 
torchvision 
torchaudio
scikit-learn
reportlab
flask-cors
google-generativeai
python-dotenv